backend/requirements.txt
fastapi
uvicorn
google-genai
google-cloud-firestore
python-dotenv
pydantic
pyvirtualdisplay
Pillow
matplotlib
numpy
graphviz
